+++ test/example/c-api-kitchen-sink.c
@@ -2161,7 +2161,9 @@ void test_typebuilder() {
                              fieldMutables,
                              2);
   }
-  TypeBuilderSetSubType(builder, tempSubStructIndex, tempStructIndex);
+  // the following two setSubType calls are identical
+  TypeBuilderSetSubTypeAt(builder, tempSubStructIndex, tempStructIndex);
+  TypeBuilderSetSubType(builder, tempSubStructIndex, tempStructHeapType);
 
   // TODO: Rtts (post-MVP?)